requirements.txt
Flask==3.0.0
ultralytics==8.3.93
numpy==2.1.1
Pillow==11.1.0
opencv-python==4.11.0.86
